% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafflemap.R
\name{waffle}
\alias{waffle}
\title{Waffle}
\usage{
waffle(v, ncol, xy, size)
}
\arguments{
\item{v}{a named vector of numeric values. It indicates the number of cells for each modalities.}

\item{ncol}{number of columns.}

\item{xy}{coordinates of the center of the waffle.}

\item{size}{size of a cell side.}
}
\value{
An sf object of polygons is returned.
}
\description{
Create a waffle.
}
\examples{
wafcells <- c(A = 2, B = 3, C = 7, D = 6, E = 4)
mywaf <- waffle(v = wafcells, ncol = 5, xy = c(0,0), size = 2)
plot(mywaf)
}
