% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafflemap.R
\name{wafflemap}
\alias{wafflemap}
\title{Waffle Map}
\usage{
wafflemap(x, var, ncol, size, ...)
}
\arguments{
\item{x}{either an sf object or a data.frame with coordinates fields named "X" and "Y".}

\item{var}{names of the field to use in x.}

\item{ncol}{number of columns.}

\item{size}{size of a cell side.}

\item{...}{other arguments from cartography::typoLayer.}
}
\value{
An sf object of polygons is returned.
}
\description{
Plot a waffle map.
}
\examples{
library(wafflemap)
library(cartography)
data(bond)
plot(st_geometry(bond[bond$ISO2 \%in\% c("IS", "TR"), ]), bg = "ivory3", border = NA)
plot(bond$geometry, border = "ivory2", bg = "ivory3", col = "ivory1", add=TRUE)
wafflemap(x = bond, var = c("Sean.Connery", "Roger.Moore", "Pierce.Brosnan",  "Daniel.Craig"),
          ncol = 3, size = 100000, add=TRUE, col = carto.pal(pal1 = "multi.pal", n1 = 4))
}
