% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafflemap.R
\name{wafflemap}
\alias{wafflemap}
\title{Waffle Map}
\usage{
wafflemap(x, var, ncol, size, ...)
}
\arguments{
\item{x}{either an sf object or a data.frame with coordinates fields named "X" and "Y".}

\item{var}{names of the field to use in x.}

\item{ncol}{number of columns.}

\item{size}{size of a cell side.}

\item{...}{other arguments from cartography::typoLayer.}
}
\value{
An sf object of polygons is returned.
}
\description{
Plot a waffle map.
}
\examples{
wafcells <- c(A = 2, B = 3, C = 7, D = 6, E = 4)
mywaf <- waffle(v = wafcells, ncol = 5, xy = c(0,0), size = 2)
plot(mywaf)
}
